pyproject.toml
[project]
name = "mem0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mem0ai>=1.0.2",
    "numpy>=2.4.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "torch",
    "torchvision",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.black]
line-length = 120
target-version = ["py313"]

